Included the Pako ZLIB library in order to reduce the size of exported EDS files.
When exporting, data is no longer dumped from memory directly in json (huge) but entropy coded using Pako and
then further encoded in base-64. The final EDS file has a 10-byte header "EDS0010000" followed by the encoded data.

When importing, the imported file is checked for the 3-byte leading "EDS". If it is found, the base-64 and entropy coding
is revesed. If it is not found, the file is already assumed in json format and used as such without inflation.

The License file was adapted to include a reference to Pako (MIT license) and ZLIB (ZLIB license).
A reference to the minified Pako library was referenced from index.html.